requirements.txt
django==5.1.4
pandas==2.2.3
flake8==7.1.1
django-import-export==4.3.3
